python-api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
supabase>=2.0
strawberry-graphql>=0.200.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
spyne>=2.14.0
lxml>=4.9.0
a2wsgi>=1.10.0
requests>=2.31.0
locust  
pandas 
matplotlib 
seaborn
